# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MeowWorld API",
        "description": "Documentación de la API para gestionar la red social.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cats": {
            "get": {
                "tags": [
                    "Gatos"
                ],
                "summary": "Obtener lista de gatos",
                "operationId": "fe45a58829a0320b90a9c6175b523284",
                "responses": {
                    "200": {
                        "description": "Lista de gatos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cat"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Gatos"
                ],
                "summary": "Crear un nuevo gato",
                "operationId": "595d23c47f7c925a8318133707a7439a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cat"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gato creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cats/filter": {
            "post": {
                "tags": [
                    "Gatos"
                ],
                "summary": "Filtrar gatos por nombre",
                "operationId": "2a99dc26514afa84515b19a8f077aa65",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "example": "Fluffy"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de gatos filtrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cat"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{userId}/cats": {
            "get": {
                "tags": [
                    "Gatos"
                ],
                "summary": "Obtener gatos de un usuario",
                "operationId": "6dc05c7c5fb03ec4c29b7adcf8130215",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gatos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cat"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cats/{id}": {
            "get": {
                "tags": [
                    "Gatos"
                ],
                "summary": "Mostrar un gato",
                "operationId": "73c5e4ac3dde88c4b28d571cd5f52a5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del gato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gato encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Gatos"
                ],
                "summary": "Actualizar un gato",
                "operationId": "9aee8f27f84e94a7f42950d36e40522c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del gato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cat"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gato actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Gatos"
                ],
                "summary": "Eliminar un gato",
                "operationId": "8cf29c418dcd721fdbf17f7089fb6d85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del gato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gato eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gato eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/filter": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Filtra usuarios por nombre de usuario",
                "operationId": "filterUsername",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Texto para buscar en los nombres de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuarios encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron usuarios"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cat": {
                "title": "Cat",
                "description": "Modelo que representa un gato en la aplicación.",
                "required": [
                    "name",
                    "description",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del gato",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del gato",
                        "type": "string",
                        "example": "Fluffy"
                    },
                    "description": {
                        "description": "Descripción del gato",
                        "type": "string",
                        "example": "Gato juguetón y cariñoso"
                    },
                    "image": {
                        "description": "URL de la imagen del gato",
                        "type": "string",
                        "example": "https://res.cloudinary.com/yourapp/cats/cat1.jpg"
                    },
                    "en_adopcion": {
                        "description": "Indica si el gato está disponible para adopción",
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "description": "ID del usuario propietario del gato",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "password",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "ID del usuario",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string"
                    },
                    "username": {
                        "description": "Nombre de usuario único",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Fecha de verificación del email",
                        "type": "string",
                        "format": "date-time"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    },
                    "description": {
                        "description": "Descripción del usuario",
                        "type": "string"
                    },
                    "img_profile": {
                        "description": "URL de la imagen de perfil",
                        "type": "string",
                        "format": "uri"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
    "security": [
        {
            "bearer": []
        }
    ]
}